requirements.txt
# Core async runtime
uvloop==0.19.0
# asyncio est intégré à Python 3.12+ (ne pas installer via pip)

# Audio processing
pydub==0.25.1
numpy==1.26.4

# AI APIs
httpx<0.28.0  # httpx 0.28+ pour compatibilité
deepgram-sdk==3.7.3
websockets<14.0  # Required for Deepgram SDK compatibility (extra_headers → additional_headers in 14.0)
groq==0.13.0
elevenlabs==1.13.1

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0

# Logging (optional, for structured logs)
structlog==24.4.0

# Database (PostgreSQL async)
asyncpg==0.29.0

# Asterisk AMI (pour récupération CALLERID via variables de canal)
panoramisk==1.4

# Web Dashboard
quart==0.19.4
jinja2==3.1.3

# Monitoring & Metrics
prometheus-client==0.20.0

# Note: FFmpeg doit être installé sur le système
# sudo apt-get install ffmpeg
